Enable Docker builds, testing, and merging with GPPD.
Major changes are in this update, mostly related to establishing a basic
workflow for updating the GPPD with estimated generation values. See the
README for the step-by-step workflow that can now be ran by anyone.

Add Dockerfile to enable building the API container, which can be ran
locally (as the project is currently configured) or on a remote machine
with some changes to the python scripts (tests/test.py,
models/estimate_*.py).
Add a couple of one-liner scripts to bring the docker container up or
down on the local machine.
Packaging this into Docker is necessary because the models rely on
specific software library versions (see Dockerfile) which are not
readily available through pip or the usual system-level package tools.
Conda is used here to match the original development environment.

Add a Makefile to provide some standardization for obtaining
non-git-controlled assets (GPPD, HydroBasins, hydro runoff data) from
remote sources.
Eventually it may make sense to amend the Makefile with the estimation
and GPPD merging process, but for now those will need to be run one at a
time with python (see README).
The Makefile also has a target for running tests.

Add tests, which currently are not using a framework (would be good to
switch to that in the future).
Tests are fairly basic right now, but check for project file system
configuration and ensure the API is returning expected values for
estimation (very non-exhaustive right now).

Add scripts to perform estimates for real GPPD plants.
These scripts produce output files that should not need to be updated
unless the underlying GPPD data meaningfully changes.
The scripts are not handling that behavior right now, but it should be a
priority in the near-future.

Add a script to merge the estimation output files with a "bare" GPPD.
The merged file is not added to git at this time.

Expand the set of fuels for which baseline estimates are available.
Add an override file to either null-out or supply a more reasonable
capacity factor for country-fuel-year combinations that were
unreasonable due to incompatibilities between the source of generation
and the source of total capacity (deficiencies, different semantics).

Convert the hydro runoff data acquisition scripts from notebooks to
plain python scripts.
This improves the ability to run this procedure from the Makefile or
CLI.
Remove the Copernicus API key (still in git history :shrugs:) and have
it supplied as a run time parameter/argument.
It is necessary to modify the Makefile to define the CDS key.

Update the gppd_ai4earth module with small formatting differences, no
major changes in function.

Add basic files to enable the API to be dockerized.

 Changes to be committed:
	modified:   README.md

	new file:   Dockerfile
	new file:   dockerdown.sh
	new file:   dockerup.sh

	new file:   Makefile

	new file:   tests/tests.py
	new file:   tests/example-gas.json
	new file:   tests/example-hydro.json
	new file:   tests/example-solar-with-age.json
	new file:   tests/example-solar-without-age.json
	new file:   tests/example-wind.json

	new file:   models/estimate_baseline.py
	new file:   models/estimate_gas.py
	new file:   models/estimate_hydro.py
	new file:   models/estimate_solar.py
	new file:   models/estimate_wind.py

	new file:   models/output/baseline_estimates.csv
	new file:   models/output/hydro_estimates.csv
	new file:   models/output/solar_estimates.csv
	new file:   models/output/wind_estimates.csv

	new file:   models/global_power_plant_database.csv
	new file:   models/make_gppd.py

	modified:   models/source_files/country_capacity_factor_preprocessing_script.py
	new file:   models/source_files/preprocessed_capacity_factors.json
	new file:   models/source_files/capacity_factor_override.json

	new file:   models/source_files/fuel_thesaurus_platts.csv
	modified:   models/source_files/gen_by_year_country_fuel.json

	deleted:    models/source_files/hydro_runoff_tool_scripts/Download_Era5_Monthly_Avg_Data.ipynb
	new file:   models/source_files/hydro_runoff_tool_scripts/download_ERA5_monthly_averages.py
	deleted:    models/source_files/hydro_runoff_tool_scripts/Transform_Era5_monthly_avg_dataset.ipynb
	new file:   models/source_files/hydro_runoff_tool_scripts/transform_ERA5_monthly_averages.py

	modified:   gppd_ai4earth/__init__.py
	modified:   gppd_ai4earth/capacity_factor_retriever.py
	modified:   gppd_ai4earth/gppd_gen/measurement_files_loader.py
	modified:   gppd_ai4earth/gppd_gen/model_runner.py
	modified:   gppd_ai4earth/sources.py

	new file:   my_api/runserver.py
	new file:   startup.sh
	new file:   supervisord.conf

	new file:   requirements.txt