Add final methods to produce GPPD v1.3.0.
Final set of scripts and processes for producing estimates and updating
a csv of the global power plant database.

This update does not make significant changes compared to the previous
commit which established the workflow.
These changes are mostly to bring the already-calculated plant-level
estimates in line with changes to the GPPD around version 1.2.4.

Changes to be committed:
	modified:   models/estimate_baseline.py
	modified:   models/estimate_hydro.py
	modified:   models/estimate_solar.py
	modified:   models/estimate_wind.py
	modified:   models/make_gppd.py
	modified:   models/output/baseline_estimates.csv
	modified:   models/output/hydro_estimates.csv
	modified:   models/output/solar_estimates.csv
	modified:   models/output/wind_estimates.csv